aider: Add the cal-heatmap script to the `manifest.json`.
# Aider chat conversation:

USER: Also add the cal-heatmap script to the `manifest.json`

ASSISTANT: To add the `cal-heatmap` script to the `manifest.json` Content Security Policy, you need to include the source of the script in the `script-src` directive. Here is the diff for the required change:

```diff
--- manifest.json
+++ manifest.json
@@ ... @@
   "content_security_policy": "script-src 'self' https://d3js.org; object-src 'self'",
+  "content_security_policy": "script-src 'self' https://d3js.org https://unpkg.com; object-src 'self'",
 }
```

This change will update the CSP to allow scripts from `https://unpkg.com` to be loaded, in addition to the existing sources.